% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.eblupres}
\alias{AIC.eblupres}
\alias{BIC.eblupres}
\title{Akaike's An Information Criterion.}
\usage{
\method{AIC}{eblupres}(object, ...)

\method{BIC}{eblupres}(object, ...)
}
\arguments{
\item{object}{EBLUP model.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
AIC value.
}
\description{
Generic function calculating Akaike's "An Information Criterion" for EBLUP model
}
\examples{
m1 <- eblupfh_cluster(y ~ x1 + x2 + x3, data = mys, vardir = "var", cluster = "clust")
AIC(m1)

}
