% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.eblupres}
\alias{autoplot.eblupres}
\title{Autoplot EBLUP results.}
\usage{
\method{autoplot}{eblupres}(object, variable = "RSE", ...)
}
\arguments{
\item{object}{EBLUP model.}

\item{variable}{variable to plot.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
plot.
}
\description{
Autoplot EBLUP results.
}
\examples{
library(saens)

m1 <- eblupfh_cluster(y ~ x1 + x2 + x3, data = mys, vardir = "var", cluster = "clust")
autoplot(m1)

}
