% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eblupfh_benchmark.R
\name{eblupfh_benchmark}
\alias{eblupfh_benchmark}
\title{Benchmarking Procedure for EBLUPs based on a Fay-Herriot Model.}
\usage{
eblupfh_benchmark(formula, vardir, method, weight, data, ...)
}
\arguments{
\item{formula}{An object of class formula that contains a description of the model to be fitted. The variables included in the formula must be contained in the data.}

\item{vardir}{Vector or column names from data that contain variance sampling from the direct estimator for each area}

\item{method}{The benchmarking method to be specified among "Ratio", "Difference", "Optimal_vardir", "Optimal_mse", "Optimal_covariance", and "You-Rao" see details.}

\item{weight}{Vector or column names from data that contain areas weights, in case of proportions it denotes the population shares.}

\item{data}{A data frame or a data frame extension (e.g. a tibble)}

\item{...}{Further arguments passed to \code{eblupfh}.}
}
\value{
The function produces the same result in both the \code{eblupfh} and \code{eblupfh_cluster} functions, returning a list containing the objects (\code{df_res} and \code{fit}). However, in \code{df_res}, additional columns for benchmark estimates and relative standard errors (RSE) are included as part of the benchmark estimation output.
}
\description{
Benchmarking Procedure for EBLUPs based on a Fay-Herriot Model.
}
\details{
The function allows performing four different benchmarking methods, according to the argument method.
\enumerate{
\item The Ratio Benchmarking (\code{Ratio}) method is one of the simplest benchmarking methods. The estimate is obtained by multiplying the EBLUP estimate with a common adjustment factor \deqn{\frac{\sum_{l=1}^{m} W_{l} \hat{\theta}_{l}}{\sum_{l=1}^{m} W_{l} \hat{\theta}_{Hl}}}
The Ratio Benchmarking method has several limitations: (i) a common adjustment factor is applied to all estimators, regardless of their precision. (ii) Additionally, the estimation from ratio benchmarking is not design-consistent, as the sample size in area-i increases while the sample sizes in the remaining areas are held fixed. (iii) Furthermore, second-order unbiased MSE estimators are not readily available (Pfeffermann, Sikov, and Tiller, 2014).
\item The Difference Benchmarking (\code{Difference}) method.
The Difference Benchmarking estimate is obtained through the following formula \deqn{\hat{\theta}^{DB}_{i} = \hat{\theta}^{H}_{i} + \left( \sum_{l=1}^{m} W_l \hat{\theta}_l - \sum_{l=1}^{m} W_l \hat{\theta}^{H}_{l} \right)}
The Difference Benchmarking method has some limitations similar to those of ratio benchmarking; however, it permits second-order unbiased MSE estimation.
\item The Optimal Benchmarking or Wang, Fuller, and Qu (WFQ Benchmarking) (\code{Optimal}) method does not suffer from the limitations (i) and (ii). The optimal benchmarking estimate is obtained through the following matrix formula
\deqn{\tilde{\theta}^{OMB} = \tilde{\theta}^H + \Omega^{-1} W \left( W^T \Omega^{-1} W \right)^{-1} \left( \hat{\theta}^+ - W^T \tilde{\theta}^H \right)}
This estimator depends on the choice of \eqn{\Omega}, popular choices of \eqn{\Omega^{-1}} include (i) \eqn{\Omega^{-1} = diag(\psi_1, \dots, \psi_m)} (\code{vardir}), (ii) \eqn{\Omega^{-1} = MSE(\hat{\theta}^H)} (\code{mse}), and (iii) \eqn{\Omega^{-1} = W^{-1} Cov(\hat{\theta}^{H}, W\hat{\theta}^{H})} (\code{covariance}).
\item The You-Rao Benchmarking (\code{Self} or \code{You-Rao}) method. The two methods are self-benchmarking (or self-calibrating) in the sense that adjustments are not made to EBLUP estimators; instead, estimators that automatically satisfy the benchmarking constraints are obtained.
Method 1 replaces the optimal \eqn{\hat{\beta}} in the EBLUP estimator \eqn{\hat{\theta}^{H}_{i}} with an estimator that depends on the benchmarking weights \eqn{W_l}.
Method 2, on the other hand, augments the vector \eqn{z_i^T} of covariates to \eqn{(z_i^T, W_i \psi_i)} and uses the EBLUP estimator of \eqn{\theta_i} based
on the augmented model. This two methods does not suffer from the limitations (i)-(iii)
}
}
\examples{
library(saens)

df_mys <- na.omit(mys)

model_rb <- eblupfh_benchmark(
  y ~ x1 + x2 + x3, data = df_mys,
  vardir = "var", weight = "weight",
  method = "ratio"
)
model_db <- eblupfh_benchmark(
  y ~ x1 + x2 + x3, data = df_mys,
  vardir = "var", weight = "weight",
  method = "Difference"
)
model_wfq <- eblupfh_benchmark(
  y ~ x1 + x2 + x3, data = df_mys,
  vardir = "var", weight = "weight",
  method = "Optimal_vardir"
)
model_yr <- eblupfh_benchmark(
  y ~ x2 + x3, data = df_mys,
  vardir = "var", weight = "weight",
  method = "You-Rao"
)
}
\references{
\enumerate{
\item Rao & Molina (2015). Small area estimation. John Wiley & Sons.
}
}
