% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.eblupres}
\alias{logLik.eblupres}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{eblupres}(object, ...)
}
\arguments{
\item{object}{EBLUP model}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Log-Likehood value
}
\description{
Extract Log-Likelihood
}
\examples{
library(saens)

model1 <- eblupfh_cluster(y ~ x1 + x2 + x3, data = mys, vardir = "var", cluster = "clust")
logLik(model1)

}
