% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mys}
\alias{mys}
\title{mys: mean years of schooling people with disabilities in Papua Island, Indonesia.}
\format{
A data frame with 42 rows and 7 variables with 10 domains are non-sampled areas.

\describe{
  \item{area}{regency municipality}
  \item{y}{mean years of schooling people with disabilities}
  \item{var}{variance sampling from the direct estimator for each area}
  \item{rse}{relative standard error (\%)}
  \item{x1}{Number of Elementary Schools}
  \item{x2}{Number of Junior High Schools}
  \item{x3}{Number of Senior High Schools}
  \item{clust}{Cluster}
  \item{n}{Number of eligible samples}
  \item{weight}{Weight}
}
}
\source{
\url{https://www.bps.go.id}
}
\usage{
mys
}
\description{
A dataset containing the mean years of schooling people with disabilities in Papua Island, Indonesia in 2021.
}
\keyword{datasets}
