% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saens.R
\name{saens}
\alias{saens}
\title{saens : Small Area Estimation (sae) for Non-sampled Areas.}
\description{
saens package Provides several functions for area level of small area estimation for non sampled areas using cluster information.
}
\references{
\enumerate{
 \item Rao, J. N., & Molina, I. (2015). Small area estimation. John Wiley & Sons.
 \item Anisa, R., Kurnia, A., & Indahwati, I. (2013). Cluster information of non-sampled area in small area estimation. E-Prosiding Internasional| Departemen Statistika FMIPA Universitas Padjadjaran, 1(1), 69-76.
}
}
\author{
Ridson Al Farizal P
}
